Began implementing MVVM design pattern Moved fragments to ui package Removed fragments package Added ViewModel classes for each fragment
