+++ app/src/main/res/layout/fragment_recipe.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.RecipeFragment">
+    tools:context=".ui.recipe.RecipeFragment">
 
     <RelativeLayout
         android:layout_width="match_parent"
