+++ app/src/main/res/layout/fragment_recipe_list.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.wit.recipes.fragments.RecipeListFragment">
+    tools:context="org.wit.recipes.ui.recipeList.RecipeListFragment">
 
 
     <androidx.recyclerview.widget.RecyclerView
