+++ app/src/main/res/layout/fragment_view_recipe.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.ViewRecipeFragment">
+    tools:context=".ui.viewRecipe.ViewRecipeFragment">
 
     <RelativeLayout
         android:layout_width="match_parent"
