+++ app/src/main/res/navigation/main_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/recipeFragment"
-        android:name="org.wit.recipes.fragments.RecipeFragment"
+        android:name="org.wit.recipes.ui.recipe.RecipeFragment"
         android:label="fragment_recipe"
         tools:layout="@layout/fragment_recipe" >
         <action
@@ -18,7 +18,7 @@
     </fragment>
     <fragment
         android:id="@+id/recipeListFragment"
-        android:name="org.wit.recipes.fragments.RecipeListFragment"
+        android:name="org.wit.recipes.ui.recipeList.RecipeListFragment"
         android:label="fragment_recipe_list"
         tools:layout="@layout/fragment_recipe_list" >
         <action
@@ -29,7 +29,7 @@
     </fragment>
     <fragment
         android:id="@+id/mapFragment"
-        android:name="org.wit.recipes.fragments.MapFragment"
+        android:name="org.wit.recipes.ui.map.MapFragment"
         android:label="fragment_map"
         tools:layout="@layout/fragment_map" />
     <fragment
